pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "a fastapi blog"
authors = ["Fernando Ike"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.3.22"
alembic = "^1.4.3"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
psycopg2 = "^2.8.6"
opentelemetry-instrumentation-fastapi = "^0.17b0"
opentelemetry-exporter-jaeger = "^0.17b0"
opentelemetry-sdk = "^0.17b0"
opentelemetry-api = "^0.17b0"
opentelemetry-instrumentation-psycopg2 = "^0.17b0"
opentelemetry-instrumentation-starlette = "^0.17b0"
opentelemetry-instrumentation-sqlalchemy = "^0.17b0"
opentelemetry-instrumentation-requests = "^0.17b0"
fastapi-pagination = "^0.5.2"
python-slugify = "^4.0.1"
passlib = "^1.7.4"
python-jose = "^3.2.0"
python-multipart = "^0.0.5"


[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
coverage = "^5.3.1"
pytest = "^6.2.1"
pytest-order = "^0.9.5"
pylint = "^2.6.0"
flake8 = "^3.8.4"
requests = "^2.25.1"
black = "^20.8b1"
pre-commit = "^2.8.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[tool.isort]
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_third_party = ["alembic", "fastapi", "fastapi_pagination", "jose", "opentelemetry", "passlib", "pydantic", "pytest", "requests", "slugify", "sqlalchemy", "starlette"]